Fix logging (#7)
* Give each pool item its own subscriber

* Add thread id to function to get thread subscriber

* Remove GuardDrop as no longer needed

* Change PoolItem trait to (hopefully) eventually provide more performant and functional tracing

* Rename re-jig and make blocking writer

* Changes to compare performance

* Remove use of global_test_scope, causing weird deadlocking problem

* re-export id based blocking

* Change behaviour of IdBasedBlocking to panic on io errors

* Add new with targets

* Add comments and tidy-up

* Simplify IdBasedBlocking new

* Move to try_lock in id_based_blocking

* Change to use an unsafecell in the IdBasedWriter

* Change to remove reload and level and dynamic levels seem slow

* Add thread id to trace

* Log dropping of IdBasedBlocking

* Change to add post processing hook. Move hooks to correct place

* Add filename formatter function to IdBasedWriter

* Make id based_writer public

* Change id back from usize to u64 so that it is consistent over 64 and 32 environments

* Try adding a macro to reduce some of the boiler plate code

* Revert "Try adding a macro to reduce some of the boiler plate code"

This reverts commit 880e0e15faf446e21d15ab7137c262edb1421982.

* Pre-release tidy up

+++ src/id_being_processed.rs
@@ -2,6 +2,6 @@ use crate::ID_BEING_PROCESSED;
 
 /// This function can be used externally to access the id of the pool item that is currently
 /// being processed
-pub fn id_being_processed() -> Option<usize> {
+pub fn id_being_processed() -> Option<u64> {
     ID_BEING_PROCESSED.with(|id_being_processed| *id_being_processed.borrow())
 }
