+++ src/samples/randoms/randoms_api/mean_request.rs
@@ -3,11 +3,11 @@ use crate::{samples::Randoms, *};
 
 /// This defines a request to calculate the mean of the contained randoms
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub struct MeanRequest(pub usize);
+pub struct MeanRequest(pub u64);
 
 /// all requests must be able to provide the id of the pool item that they are targeting
 impl IdTargeted for MeanRequest {
-    fn id(&self) -> usize {
+    fn id(&self) -> u64 {
         self.0
     }
 }
