+++ src/samples/randoms/randoms_api/panic_request.rs
@@ -3,11 +3,11 @@ use crate::{samples::Randoms, *};
 
 /// This defines a request to calculate the mean of the contained randoms
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PanicRequest(pub usize);
+pub struct PanicRequest(pub u64);
 
 /// all requests must be able to provide the id of the pool item that they are targeting
 impl IdTargeted for PanicRequest {
-    fn id(&self) -> usize {
+    fn id(&self) -> u64 {
         self.0
     }
 }
