+++ src/samples/randoms/randoms_api/randoms_add_request.rs
@@ -2,10 +2,10 @@ use crate::{samples::Randoms, *};
 
 /// This is message that sent to request the creation of a new Randoms struct with the specified id
 #[derive(Debug, PartialEq, Eq, Clone)]
-pub struct RandomsAddRequest(pub usize);
+pub struct RandomsAddRequest(pub u64);
 
 impl IdTargeted for RandomsAddRequest {
-    fn id(&self) -> usize {
+    fn id(&self) -> u64 {
         self.0
     }
 }
