+++ src/samples/randoms_batch/randoms_batch_api/sum_of_sums_request.rs
@@ -9,10 +9,10 @@ use super::{RandomsBatchApi, SumOfSumsResponse};
 /// This is the message that is sent to request that a given RandomsBatch calculates the sum of all of the
 /// sums of its contained Randoms
 #[derive(Debug, PartialEq, Eq, Clone)]
-pub struct SumOfSumsRequest(pub usize);
+pub struct SumOfSumsRequest(pub u64);
 
 impl IdTargeted for SumOfSumsRequest {
-    fn id(&self) -> usize {
+    fn id(&self) -> u64 {
         self.0
     }
 }
