+++ src/thread_request_response/id.rs
@@ -4,7 +4,7 @@ impl<P> ThreadRequestResponse<P>
 where
     P: PoolItem,
 {
-    pub fn id(&self) -> usize {
+    pub fn id(&self) -> u64 {
         match self {
             ThreadRequestResponse::ThreadShutdown(request_response) => request_response.id(),
             ThreadRequestResponse::ThreadAbort(request_response) => request_response.id(),
